chore(docs/user): package Mathlib4 in Nix
Well, this is required… Following Loogle's steps, we add fake files for
ProofWidgets4 and we can produce an .olean from Aeneas' base library
without Lake involvement!

Free parallelization and caching :).

TODOs are to improve the hash handling when lake-manifest.json bumps,
but this requires extending Lake to produce the NAR hash serializations.

Signed-off-by: Ryan Lahfa <ryan.lahfa@inria.fr>