chore(lean): minimize imports in `Base/Arith`
This is a methodology employed by the Mathlib community to improve the
build performance of Mathlib.

The idea is to exploit the ImportGraph library which has a command
`#min_imports` and `#redundant_imports` to figure out what imports to
trim down.

See: https://reservoir.lean-lang.org/@leanprover-community/importGraph
for the docs.

This is the result of this process for `Base/Arith`.

Signed-off-by: Ryan Lahfa <ryan.lahfa@inria.fr>